Moved the Input class into it's own file
In preparation for adding more pre-processing. Specifically reading in mesh files to obtain the coordinates of nodes.
Edited imports to match.
